+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="co.netguru.android.template"
+<manifest package="co.netguru.baby.monitor.client"
     xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
