--- app/src/main/kotlin/co/netguru/android/template/application/App.kt
@@ -1,24 +1,0 @@
-package co.netguru.android.template.application
-
-import dagger.android.AndroidInjector
-import dagger.android.support.DaggerApplication
-import io.reactivex.plugins.RxJavaPlugins
-import javax.inject.Inject
-
-class App : DaggerApplication() {
-
-    @Inject
-    lateinit var debugMetricsHelper: DebugMetricsHelper
-
-    @Inject
-    lateinit var rxJavaErrorHandler: RxJavaErrorHandler
-
-    override fun onCreate() {
-        super.onCreate()
-        debugMetricsHelper.init(this)
-        RxJavaPlugins.setErrorHandler(rxJavaErrorHandler)
-    }
-
-    override fun applicationInjector(): AndroidInjector<App> =
-            DaggerApplicationComponent.builder().create(this)
-}
