--- app/src/main/kotlin/co/netguru/android/template/application/ApplicationComponent.kt
@@ -1,20 +1,0 @@
-package co.netguru.android.template.application
-
-import co.netguru.android.template.application.scope.AppScope
-import dagger.Component
-import dagger.android.AndroidInjectionModule
-import dagger.android.AndroidInjector
-import dagger.android.support.AndroidSupportInjectionModule
-
-@AppScope
-@Component(
-        modules = [
-            AndroidInjectionModule::class,
-            AndroidSupportInjectionModule::class,
-            ApplicationModule::class
-        ]
-)
-internal interface ApplicationComponent : AndroidInjector<App> {
-    @Component.Builder
-    abstract class Builder : AndroidInjector.Builder<App>()
-}
