--- app/src/main/kotlin/co/netguru/android/template/application/ApplicationModule.kt
@@ -1,13 +1,0 @@
-package co.netguru.android.template.application
-
-import co.netguru.android.template.application.scope.AppScope
-import dagger.Module
-import dagger.Provides
-
-@Module
-class ApplicationModule {
-
-    @AppScope
-    @Provides
-    fun rxJavaErrorHandler(): RxJavaErrorHandler = RxJavaErrorHandlerImpl()
-}
