--- app/src/main/kotlin/co/netguru/android/template/application/scope/AppScope.kt
@@ -1,6 +1,0 @@
-package co.netguru.android.template.application.scope
-
-import javax.inject.Scope
-
-@Scope
-annotation class AppScope
