--- app/src/main/kotlin/co/netguru/android/template/common/extensions/SharedPreferencesExtensions.kt
@@ -1,9 +1,0 @@
-package co.netguru.android.template.common.extensions
-
-import android.content.SharedPreferences
-
-inline fun SharedPreferences.edit(action: SharedPreferences.Editor.() -> Unit) {
-    val editor = edit()
-    editor.action()
-    editor.apply()
-}
