--- app/src/main/kotlin/co/netguru/android/template/feature/splash/SplashActivity.kt
@@ -1,5 +1,0 @@
-package co.netguru.android.template.feature.splash
-
-import android.support.v7.app.AppCompatActivity
-
-class SplashActivity : AppCompatActivity()
