+++ app/src/main/kotlin/co/netguru/baby/monitor/client/application/ApplicationComponent.kt
@@ -1,0 +1,20 @@
+package co.netguru.baby.monitor.client.application
+
+import co.netguru.baby.monitor.client.application.scope.AppScope
+import dagger.Component
+import dagger.android.AndroidInjectionModule
+import dagger.android.AndroidInjector
+import dagger.android.support.AndroidSupportInjectionModule
+
+@AppScope
+@Component(
+        modules = [
+            AndroidInjectionModule::class,
+            AndroidSupportInjectionModule::class,
+            ApplicationModule::class
+        ]
+)
+internal interface ApplicationComponent : AndroidInjector<App> {
+    @Component.Builder
+    abstract class Builder : AndroidInjector.Builder<App>()
+}
