+++ app/src/main/kotlin/co/netguru/baby/monitor/client/application/scope/FragmentScope.kt
@@ -1,0 +1,6 @@
+package co.netguru.baby.monitor.client.application.scope
+
+import javax.inject.Scope
+
+@Scope
+annotation class FragmentScope
