+++ app/src/main/kotlin/co/netguru/baby/monitor/client/common/extensions/SharedPreferencesExtensions.kt
@@ -1,0 +1,9 @@
+package co.netguru.baby.monitor.client.common.extensions
+
+import android.content.SharedPreferences
+
+inline fun SharedPreferences.edit(action: SharedPreferences.Editor.() -> Unit) {
+    val editor = edit()
+    editor.action()
+    editor.apply()
+}
