+++ app/src/main/kotlin/co/netguru/baby/monitor/client/feature/splash/SplashActivity.kt
@@ -1,0 +1,5 @@
+package co.netguru.baby.monitor.client.feature.splash
+
+import android.support.v7.app.AppCompatActivity
+
+class SplashActivity : AppCompatActivity()
