--- app/src/test/kotlin/co/netguru/android/template/RxSchedulersOverrideRule.kt
@@ -1,28 +1,0 @@
-package co.netguru.android.template
-
-import io.reactivex.android.plugins.RxAndroidPlugins
-import io.reactivex.plugins.RxJavaPlugins
-import io.reactivex.schedulers.Schedulers
-import org.junit.rules.TestRule
-import org.junit.runner.Description
-import org.junit.runners.model.Statement
-
-class RxSchedulersOverrideRule : TestRule {
-
-    override fun apply(base: Statement, description: Description): Statement {
-        return object : Statement() {
-            @Throws(Throwable::class)
-            override fun evaluate() {
-                RxAndroidPlugins.setInitMainThreadSchedulerHandler { Schedulers.trampoline() }
-                RxJavaPlugins.setComputationSchedulerHandler { Schedulers.trampoline() }
-                RxJavaPlugins.setIoSchedulerHandler { Schedulers.trampoline() }
-                RxJavaPlugins.setNewThreadSchedulerHandler { Schedulers.trampoline() }
-                RxJavaPlugins.setSingleSchedulerHandler { Schedulers.trampoline() }
-                base.evaluate()
-
-                RxJavaPlugins.reset()
-                RxAndroidPlugins.reset()
-            }
-        }
-    }
-}
