+++ app/src/test/kotlin/co/netguru/baby/monitor/RxTestSchedulerOverrideRule.kt
@@ -1,0 +1,35 @@
+package co.netguru.baby.monitor
+
+import io.reactivex.android.plugins.RxAndroidPlugins
+import io.reactivex.plugins.RxJavaPlugins
+import io.reactivex.schedulers.Schedulers
+import io.reactivex.schedulers.TestScheduler
+import org.junit.rules.TestRule
+import org.junit.runner.Description
+import org.junit.runners.model.Statement
+
+/**
+ * Rule helpful while testing blocking/waiting rxjava code.
+ * All handlers are set to main thread, Computation scheduler is replaced with [TestScheduler]
+ * accessible through [testScheduler] property.
+ */
+class RxTestSchedulerOverrideRule : TestRule {
+    val testScheduler = TestScheduler()
+
+    override fun apply(base: Statement, description: Description): Statement {
+        return object : Statement() {
+            @Throws(Throwable::class)
+            override fun evaluate() {
+                RxAndroidPlugins.setInitMainThreadSchedulerHandler { Schedulers.trampoline() }
+                RxJavaPlugins.setComputationSchedulerHandler { testScheduler }
+                RxJavaPlugins.setIoSchedulerHandler { Schedulers.trampoline() }
+                RxJavaPlugins.setNewThreadSchedulerHandler { Schedulers.trampoline() }
+                RxJavaPlugins.setSingleSchedulerHandler { Schedulers.trampoline() }
+                base.evaluate()
+
+                RxJavaPlugins.reset()
+                RxAndroidPlugins.reset()
+            }
+        }
+    }
+}
